+++ service/collector_test.go
@@ -36,6 +36,8 @@ import (
 	"go.uber.org/zap/zapcore"
 
 	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/mapconverter/overwritepropertiesmapconverter"
 	"go.opentelemetry.io/collector/internal/testcomponents"
 	"go.opentelemetry.io/collector/internal/testutil"
 	"go.opentelemetry.io/collector/service/featuregate"
@@ -60,11 +62,18 @@ func TestCollector_StartAsGoRoutine(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+		"yaml:service::telemetry::metrics::address: localhost:" + portAsString(testutil.GetAvailablePort(t)),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	set := CollectorSettings{
-		BuildInfo: component.NewDefaultBuildInfo(),
-		Factories: factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-config.yaml")},
-			[]string{"service.telemetry.metrics.address=localhost:" + strconv.FormatUint(uint64(testutil.GetAvailablePort(t)), 10)}),
+		BuildInfo:      component.NewDefaultBuildInfo(),
+		Factories:      factories,
+		ConfigProvider: cfgProvider,
 	}
 	col, err := New(set)
 	require.NoError(t, err)
@@ -98,12 +107,23 @@ func testCollectorStartHelper(t *testing.T) {
 	}
 
 	metricsPort := testutil.GetAvailablePort(t)
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+	}
+	cfgSet.MapConverters = append([]config.MapConverterFunc{
+		overwritepropertiesmapconverter.New(
+			[]string{"service.telemetry.metrics.address=localhost:" + portAsString(metricsPort)},
+		)},
+		cfgSet.MapConverters...,
+	)
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	col, err := New(CollectorSettings{
-		BuildInfo: component.NewDefaultBuildInfo(),
-		Factories: factories,
-		ConfigProvider: MustNewDefaultConfigProvider(
-			[]string{filepath.Join("testdata", "otelcol-config.yaml")},
-			[]string{"service.telemetry.metrics.address=localhost:" + strconv.FormatUint(uint64(metricsPort), 10)}),
+		BuildInfo:      component.NewDefaultBuildInfo(),
+		Factories:      factories,
+		ConfigProvider: cfgProvider,
 		LoggingOptions: []zap.Option{zap.Hooks(hook)},
 	})
 	require.NoError(t, err)
@@ -169,10 +189,17 @@ func TestCollector_ShutdownNoop(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	set := CollectorSettings{
 		BuildInfo:      component.NewDefaultBuildInfo(),
 		Factories:      factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-config.yaml")}, nil),
+		ConfigProvider: cfgProvider,
 	}
 	col, err := New(set)
 	require.NoError(t, err)
@@ -190,10 +217,17 @@ func TestCollector_ShutdownBeforeRun(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	set := CollectorSettings{
 		BuildInfo:      component.NewDefaultBuildInfo(),
 		Factories:      factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-config.yaml")}, nil),
+		ConfigProvider: cfgProvider,
 	}
 	col, err := New(set)
 	require.NoError(t, err)
@@ -221,11 +255,18 @@ func TestCollector_ClosedStateOnStartUpError(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-invalid.yaml"),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	// Load a bad config causing startup to fail
 	set := CollectorSettings{
 		BuildInfo:      component.NewDefaultBuildInfo(),
 		Factories:      factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-invalid.yaml")}, nil),
+		ConfigProvider: cfgProvider,
 	}
 	col, err := New(set)
 	require.NoError(t, err)
@@ -256,10 +297,17 @@ func TestCollector_ReportError(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	col, err := New(CollectorSettings{
 		BuildInfo:      component.NewDefaultBuildInfo(),
 		Factories:      factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-config.yaml")}, nil),
+		ConfigProvider: cfgProvider,
 	})
 	require.NoError(t, err)
 
@@ -288,11 +336,18 @@ func TestCollector_ContextCancel(t *testing.T) {
 	factories, err := testcomponents.NewDefaultFactories()
 	require.NoError(t, err)
 
+	cfgSet := newDefaultConfigProviderSettings()
+	cfgSet.Locations = []string{
+		filepath.Join("testdata", "otelcol-config.yaml"),
+		"yaml:service::telemetry::metrics::address: localhost:" + portAsString(testutil.GetAvailablePort(t)),
+	}
+	cfgProvider, err := NewConfigProvider(cfgSet)
+	require.NoError(t, err)
+
 	set := CollectorSettings{
-		BuildInfo: component.NewDefaultBuildInfo(),
-		Factories: factories,
-		ConfigProvider: MustNewDefaultConfigProvider([]string{filepath.Join("testdata", "otelcol-config.yaml")},
-			[]string{"service.telemetry.metrics.address=localhost:" + strconv.FormatUint(uint64(testutil.GetAvailablePort(t)), 10)}),
+		BuildInfo:      component.NewDefaultBuildInfo(),
+		Factories:      factories,
+		ConfigProvider: cfgProvider,
 	}
 	col, err := New(set)
 	require.NoError(t, err)
@@ -376,4 +431,8 @@ func assertZPages(t *testing.T) {
 	for _, path := range paths {
 		testZPagePathFn(t, path)
 	}
+}
+
+func portAsString(p uint16) string {
+	return strconv.FormatUint(uint64(p), 10)
 }